Verify if nagios user exists before attempting to run a check
There are cases where a charm may generate its configuration before
nagios-nrpe-server has been installed and attempting to run a check will
cause a failure due that /var/lib/nagios does not exist.

Fixes: #1670223

+++ charmhelpers/contrib/charmsupport/nrpe.py
@@ -383,7 +383,10 @@ def add_init_service_checks(nrpe, services, unit_name, immediate_check=True):
                 description='service check {%s}' % unit_name,
                 check_cmd='check_status_file.py -f %s' % checkpath,
             )
-            if immediate_check:
+            # if nagios user doesn't exist, it means nagios-nrpe-server hasn't
+            # been installed yet and /var/lib/nagios doesn't exist either, so
+            # open(checkpath, 'w') will fail (LP: #1670223).
+            if immediate_check and host.user_exists('nagios'):
                 f = open(checkpath, 'w')
                 subprocess.call(
                     croncmd.split(),
